don't use two uuid libraries at once
pborman/uuid has been a wrapper around google/uuid for a while, so just
use google/uuid directly.

Note that in pborman/uuid, New returns a hex-encoded random UUID. In the
new API, the conversion to a hex-encoded string is explicit.

While at it, go mod tidy and gofmt -s.