allow running integration tests against an existing lotus node
The added tunnel.sh script opens an ssh proxy-forwarding tunnel for the
API, and sets up the LOTUS_API and LOTUS_TOKEN env vars for dealbot to
be able to talk to it.

The header.sh script is updated to only run a local devnet if those env
vars are unset.

Finally, remove 01_storage_retrieval_ok_local, as we no longer seem to
have issues with the first deail failing.