more slight improvements to the integration tests
After Will's changes to the flag handling, we can now set the lotus node
API address and token globally via an env var, without having to pass
that through via flags.

Do the same for data-dir and miner.

I also noticed that the "go install ./..." now trashed the cache, since
it ran inside a temporary HOME. Oops, fixed.

We also want to run "go install" even when we don't need devnet, to
ensure we're using the right dealbot binary.

Also add context to the import error, as I was hitting it and a generic
"no such file" error isn't helpful.