add sshfs to tunnel script for storage deals
This makes the existing integration test not error immediately, since
the remote machine would not have access to the file path in a local
temporary directory.

Note that, right now, the test doesn't fail but seems to just hang.
Presumably, this is because the remote test network takes a while to
publish and advance deals, similar to mainnet, whereas devnet can make a
storage deal finish in under a minute.