controller: minor code review adjustments post-merge
Make sure that controller/client's HTTP client has a global timeout, to
ensure requests don't block forever even if the caller forgets to set up
a context with a deadline.

Use buffers directly as request readers. Since they are one-off, we
don't need to wrap them in a new reader.

Be more consistent about logging and returning when the controller runs
into an error serving a request.