feat(dealbot): enable local testing with fake dealbot
Enable a full local testing node by using a deal bot that simply reports statuses without making
deals. With this mock dealbot, we can setup a docker-compose stack that will run everything up to
prometheus (along with postgres) so that you can work with all parts of the stack that are not
lotus. Also adds command lines for queue-ing storage and retrieval at the command line.