devnet: stream logs to files, fix delays
The logs used to be buffered, only printing the last few lines on error.
That's not enough if something doesn't work and we want to look at the
logs while devnet runs, though. So, instead, write to a log file in the
current directory for both the miner and the daemon.

The select to sleep for the two periodic commands was also broken: it
had a "default" clause, so we never slept for the intended 30s. We would
essentially run the command in a loop without delay, which is far too
often.

The new code actually works. We also lower the delay from 30s to 5s, to
keep the tests running relatively quickly. Running a devnet locally on
my laptop goes from using ~40% CPU on average, to about ~10% CPU.

Finally, we make the chains of shell commands into []string lists, which
makes them much more readable.