integration_tests: add first test including the controller
This test starts a controller and a daemon, adds a storage task to the
controller queue, and waits for the daemon to pick it up and complete
it.

	$ bash integration_tests/02_controller_daemon.sh
	[about five minutes later]
	daemon.log: [...] engine/engine.go:111 successfully stored data
	Storage deal is active!

This test is slow, but also pretty barebones; a number of TODOs are
included for ways in which we could extend it.

We remove "set -x" from the header script, because it would get far too
noisy with the "grep" querying of the logs until completion.

This commit also includes an improved tunnel script, to work with the new AWS
machine, plus docs on how to use it.