fix logging of tasks since we moved to IPLD
Task fields are no longer plain Go values, as they are IPLD nodes that
wrap the actual values in unexported fields. As such, we'd get pointless
log lines like:

	retrieval task  {"uuid": {}, "miner": {}, "payload_cid": {}, [...]

With this change, we go back to logging the simple Go values:

	storage task    {"uuid": "67fcf61a-...", "miner": "t01000", "size": 1024, [...]

Add a test via the second integration test, as it already runs the
controller with a real task.

While at it, be nicer with mktemp, giving a human-friendly suffix to its
directories.

Fixes #100.