Tasks can be scheduled for repeated execution
Storage and retrieval tasks have a "Schedule" string, which when not empty specifies a schedule for repeated task execution.  The tasks also have an optional  "ScheduleLimit" string which is a golang duration string (e.g. "72h3m") and specifies for how long the task will keep getting scheduled.  If a task has no schedule then it is run immediately, one time. If a task has a Schedule, but no ScheduleLimit, then the task is scheduled indefinitely.

Each time a task is executed, it is started from the same state as when it was retrieved by PopTask.