tasks: record first and last byte for retrievals
And do a better job at recording "first byte" for storage deals,
since now we have an "opening data transfer" log entry.

Note that we had to add UpdatedAt timestamps to all the static stages
recorded in global maps. We needed the timestamp from some of them for
this feature, such as RetrievalStages["FirstByteReceived"].

Unfortunately, since the IPLD nodes were computed at init time, we
didn't have a chance to set the timestamp at each time the stages are
recorded.

We can't do that in the engine package either, as the node struct's
inner fields are unexported. The next best option is to turn those maps
into a form of constructors; they now provide functions which give a
StageDetails filled with time.Now when called.

Adapting the rest of the codebase to those func calls was fairly easy.
The maps with func values are perhaps not the most elegant solution,
but certainly the smallest reasonable change I could think of.

Fixes #88.