Skip task generation if previous task incomplete (#252)
* Skip task generation if previous task incomplete

Migration creates a new parent column in the tasks table. If a task is generated from a scheduled task, then the generated task's parent is set to the ID of the scheduled task. When a scheduled task is run, first a query looks for any tasks that are children of the scheduled (generator) task, and if any are unfinished (status < Successful), then task generation is skipped.

This fixes #251