Publish Pando-compatible metadata over GraphSync
Implement a publisher specific to Pando metadata that uses the CID of
`RecordUpdate` as byte value of Pando metadata and exposes a GraphSync
server that can resolve both Pando metadata, and the `RecordUpdate`
CIDs.

Decouple `statedb` from publication by changing the state interface to
return the CID of `RecordUpdate`. This removes all libp2p and legs code
from the `state` package. Instead, the publication is done on the
controller itself upon `POST`s to `/complete/<workedby>` as a
non-critical operation. This makes a better structured code in the
context of Separation of Concerns.

Relates to:
 - https://github.com/filecoin-project/dealbot/issues/342