Fix dt datastore decode issue and set latest on publisher startup
Fix an issue where storage of bytes attempted by the datatransfer manger
fail due to UTf-8 econding at DB level. Use an ephemeral datastore for
datatransfer manager to avoid this altogether. Left a TODO to revisit.

Set the reference to root CID if available on publisher startup.

Add tests to assert publication, retrieval and chaining of metadata in
dealbot.